apt-cudf: improve handling of temporary files and fifos
The new approach create a tempdir and stores below it both the FIFO for CUDF
input and the output file for CUDF solution.

Rationale: the former approach, based on UUID generation, was prone to a race
condition (and hence potentially to a tempfile attack). The new one should also
be nicer to the user, as the temp directory can be easily recognized as
belonging to apt-cudf; it was not the case for UUIDs.