cleanup in the deprecated implementation of version (de)composition.
Remove in the deprecated part some auxiliary functions that are no longer
used. Also, some functions for splitting off different parts of a version
(epoch, etc.) have in the meanwhile been redefined for the reimplementation
of version comparison and decomposition. Make these auxiliary functions
used in the deprecated code local functions of split, to avoid
confusing with the new implementation of these functions with the
same name.