Add new function in Boilerplate
add if_application to use ml files in the application folder
as modules and test them programmatically. if application
works similarly to " if __name__ == '__main__': " in python.

This is handy because sometimes applications contain specific
logic that does not belong to the core library but that should
be unit-tested anyway.

challenged and outdated are invoked using if_application.