Various Optimizations
try to avoid repeated lookup for dependencies and conflicts
in depsolver_int, strongdeps_int and defaultgraphs

add a _cache version of init solver where we can precompute
all dependencies and conflict for a given package. this is
useful if we want to call the solver over and over again
using the same universe.

Apparently distcheck is consistently 0.6 seconds faster now.
Strongdeps is still slower then commit
75eb4a2fcee5629796ab8ed1145ddee28c1abfd2