distcheck option handling: more uniform usage of implicit resource prefix
The implicit resource prefix derived from the name of the executable now
applies to positin arguments and to values of the options --bg and --fg.
(closes: bug#348). Also, enable for eclipsecheck, rpmcheck the same rule
as for debcheck: if no position argument is given then forground is read
from stdin.