fix Strong Deps and API change
The strong dependencies application should not consider essential
debian packages. This was causing an exception as the dependency
clusure function does not consider essential packages.

The Depsolver_int.solve function now has a different output type.
Depsolver_int.solve output is in terms of solver integers and not
cudf integers.