deb-buildcheck.ml: Add cross compilation capability
Using the new --host option, one can now specify an architecture for
which packages should be attempted to be cross compiled for. The
architecture given in the --host option must naturally also be listed in
the --deb-foreign-arch option.

In contrast to native compilation, the native, foreign and host
properties of Debcudf.default_options are set to their respective values
when cross compiling.