deb/sources.ml: build-essential:native is always a build dependency
By policy, build-essential is a build dependency of every source
package. In particular, apt always adds build-essential:native to the
dependencies of every source package when resolving build dependencies.

Apt therefor fails, should a source package build depend on a foreign
version of a package that is also required by multiarch but not
Multi-Arch: same.

This patch lets dose3 behave the same way apt-get currently does.

In the future, there might be an implicit dependency on
build-essential-cross for cross building source packages.