Add multi-objective optimization routines
This patch substantaly modify the edos solver to add moop capabilities.

We add two new functions solve_pbo and solve_all . The first search
for the best solution according to the given criteria, the second
explore the entire search space. Both functions accept an additional
function as parameter (callback) that will be called each time the
best (or any) solution is found.

The function check_request in depsolver.ml was modified accordingly
to use the solve_pbo routine.