deb/packages.ml: remove build-essential field
The field was exposed since commit
f549494786c3892b715808689927aff7a463a46a in dose3 version 4.1 but the
field is not useful for any tool consuming data from dose3.

The field was recently added to dpkg as part of bug
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=806315 but neither the
reporter of the bug nor guillem have any use for the field. Manuel told
me that they only reported the bug because dpkg printed that warning and
not because they needed the field. Guillem notes in the bug that the
field has never been documented and is not even part of the respective
source packages but instead automatically injected into Packages files
by the archive. When talking with Guillem we agreed that the dpkg
support is only to document the existence of the field and not to
propagate its use for a yet unknown use case.

The right way of handling packages of the build-essential set is the
build-essential package together with its transitive dependencies. I
assume the field comes from a time when dependency resolvers where
unable to resolve dependencies reliably and thus this field got added.
Users of dose3 should not feel to be in lack of a powerful dependency
resolver though, so the field is definitely not needed in its output.