applications/apt-cudf.ml: Do not generate removal requests for up/downgrades
As far as solvers are concerned, changes in package versions mean that
the old package is removed and the package in the new version is
installed. This is how it was encoded in the EDSP output so far.

Recently, (commit c0d57a2a) the order of installation and removal
requests was turned around and it was discovered that apt will only
parse the *last* request for a package (see Debian bug #823906). Thus,
apt would remove a certain package but not install it in its new
version.

As a remedy, this commit makes sure that package up or downgrades are
expressed in a single "Install" request and that these do not show up as
a "Remove" request anymore.