applications/deb-buildcheck.ml: fix regression that lead to Essential:yes packages not being considered
Generating the sub-universe doesn't work anymore by using CudfAdd.cone
because that ignores the implicit dependencies on Essential:yes
packages. Instead, the depsolver dependency module would have to
calculate a dependency closure. At this point the added complexity is
probably not worth the benefit. We thus do not calculate a sub-universe
anymore but carry out a normal listcheck operation.