distcheck: Error out if --coinst packages cannot be found
While it might be okay to silently fail if one of the packages given in
the --checkonly vpkglist cannot be found in the universe, we want to
complain loudly if one of the packages given in the --coinst vpkglist
cannot be found. This is because in contrast to --checkonly, the
vpkglist given via --coinst only makes sense for the user if none of its
items gets ignored.

Ref [#20703] The debcheck --coinst option does not give a warning if the
specified packages cannot be found.